requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
networkx==2.8.8
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
requests==2.28.1
rich==12.6.0
scikit-image==0.19.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.14.0
tifffile==2022.10.10
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
zipp==3.10.0
